.k8s/demo/requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
requests==2.25.1
urllib3==1.26.2
pytest==6.2.4
psycopg2_binary==2.9.1
helm-python==0.1.3
kubernetes==17.17.0
toml==0.10.1
pyyaml==5.3.1
termcolor==1.1.0